% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchplot.R
\name{pitchplot}
\alias{pitchplot}
\title{Plot pitch object}
\usage{
pitchplot(
  pt,
  start,
  end,
  tfrom0 = TRUE,
  tgbool = FALSE,
  lines = NULL,
  plotType = "draw",
  scale = "hz",
  freqRange = c(50, 500),
  semitonesRe = 100,
  color = "black",
  ind = NULL,
  nframe = NULL,
  start_end_only = TRUE,
  min_max_only = TRUE,
  axisLabel = NULL
)
}
\arguments{
\item{pt}{PitchTier object loaded using \link[rPraat:pt.read]{rPraat::pt.read}}

\item{start}{Start time (in seconds) of desired plotted area.}

\item{end}{End time (in seconds) of desired plotted area.}

\item{tfrom0}{Logical; should time on the x-axis run from 0 or from the
original time? Default is \code{TRUE}.}

\item{tgbool}{Logical; should dotted lines be plotted corresponding to
locations in a TextGrid? Default is \code{FALSE}.}

\item{lines}{Numeric vector giving locations in seconds of locations from
a TextGrid to be plotted with dotted lines. Default is \code{NULL}.}

\item{plotType}{String giving the type of pitch plot to produce; default
is \code{draw} (a line plot), the only other option is \code{speckle} (a point plot).}

\item{scale}{String giving the frequency scale to use when producing
pitch plots. Default is \code{hz}; other options are \code{logarithmic} (also in Hz),
\code{semitones}, \code{erb}, and \code{mel}.}

\item{freqRange}{Vector of two integers giving the frequency range to be
used for producing pitch plots. Default is \code{c(50,500)}. If the frequency
scales \code{semitones} or \code{erb} are used, the pitch range is automatically reset
to the Praat defaults for these scales (\code{c(-12,30)} and \code{c(0,10)},
respectively).}

\item{semitonesRe}{Frequency in Hz giving the reference level for converting
pitch frequency to semitones. Default is \code{100}.}

\item{color}{String giving the name of the color to be used for
plotting pitch. Default is \code{'black'}.}

\item{ind}{Integer indexing waveform relative to other plot components.
Default is \code{NULL}.}

\item{nframe}{Integer giving the number of plot components. Default is \code{NULL}.}

\item{start_end_only}{Logical; should there only be ticks on the x-axis
for start and end times? Default is \code{TRUE}.}

\item{min_max_only}{Logical; should only minimum and maximum values be given
on the y-axis? Default is \code{TRUE}. Can also be a logical vector if some but
not all plot components should have minimum and maximum values on the y-axis.
Ignored for TextGrid component.}

\item{axisLabel}{String giving the name of the label to print along the
y-axis when printing a pitch track. Default is \code{NULL}, in which case the
axis label will depend on the scale.}
}
\description{
Function for plotting pitch objects called by \link{praatpicture}. Instead of
using this function directly, just use
\code{praatpicture('my_sound_file', frames='pitch')}.
}
\examples{
#dont use directly
datapath <- system.file('extdata', package='praatpicture')
praatpicture(paste0(datapath, '/1.wav'), frames='pitch')
}
