% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_praatpicture.R
\name{shiny_praatpicture}
\alias{shiny_praatpicture}
\title{Run praatpicture as Shiny app}
\usage{
shiny_praatpicture()
}
\value{
No return values
}
\description{
Interactive version of \link{praatpicture}
}
\examples{
\dontrun{
shiny_praatpicture()
}
}
