% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgplot.R
\name{tgplot}
\alias{tgplot}
\title{Plot TextGrid}
\usage{
tgplot(
  tg,
  t,
  sr,
  start,
  tiers = 1,
  tfrom0 = TRUE,
  tier_names = TRUE,
  ind = NULL,
  nframe = NULL
)
}
\arguments{
\item{tg}{TextGrid object loaded using \link[rPraat:tg.read]{rPraat::tg.read}}

\item{t}{Numeric vector giving times corresponding to the signal.}

\item{sr}{Integer giving the sampling rate of the signal.}

\item{start}{Start time (in seconds) of desired plotted area.}

\item{tiers}{Vector of number or strings giving either numeric identifiers
of TextGrid tiers to plot or the names of TextGrid tiers to plot. Default is
\code{1}, which plots just the first tier.}

\item{tfrom0}{Logical; should time on the x-axis run from 0 or from the
original time? Default is \code{TRUE}.}

\item{tier_names}{Logical; should TextGrid tier names be printed along the
y-axis? Default is \code{TRUE}.}

\item{ind}{Integer indexing waveform relative to other plot components.
Default is \code{NULL}.}

\item{nframe}{Integer giving the number of plot components. Default is \code{NULL}.}
}
\description{
Function for plotting TextGrids called by \link{praatpicture}. Instead of using
this function directly, just use
\code{praatpicture('my_sound_file', frames='TextGrid')}.
}
\examples{
#dont use directly
datapath <- system.file('extdata', package='praatpicture')
praatpicture(paste0(datapath, '/1.wav'), frames='TextGrid')
}
