% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv2sc.R
\name{conv2sc}
\alias{conv2sc}
\title{Convert capital letters to Unicode small caps}
\usage{
conv2sc(x)
}
\arguments{
\item{x}{A string where all capital letters should be converted to small
caps.}
}
\value{
A string where all capital letters have been converted to small caps.
}
\description{
Helper function to convert capital letters into Unicode small caps.
May not work for all font families. Note that there's no Unicode small cap
'X', so 'X' will just be converted to 'x'.
}
\examples{
my_string <- 'ABCDEFGHIJKLMNOPQRSTUVWXYZ'
conv2sc(my_string)
}
