% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formantplot.R
\name{formantplot}
\alias{formantplot}
\title{Plot formant object}
\usage{
formantplot(
  fm,
  start,
  end,
  tfrom0 = TRUE,
  tgbool = FALSE,
  lines = NULL,
  dynamicrange = 30,
  formantrange = c(0, 5500),
  type = "speckle",
  ind = NULL,
  nframe = NULL
)
}
\arguments{
\item{fm}{Formant object loaded using \link[rPraat:formant.read]{rPraat::formant.read}}

\item{start}{Start time (in seconds) of desired plotted area.}

\item{end}{End time (in seconds) of desired plotted area.}

\item{tfrom0}{Logical; should time on the x-axis run from 0 or from the
original time? Default is \code{TRUE}.}

\item{tgbool}{Logical; should dotted lines be plotted corresponding to
locations in a TextGrid? Default is \code{FALSE}.}

\item{lines}{Numeric vector giving locations in seconds of locations from
a TextGrid to be plotted with dotted lines. Default is \code{NULL}.}

\item{dynamicrange}{Dynamic range in dB for producing formant plots.
When a formant plot of \code{formanttype='speckle'} is drawn, no formants are
shown in frames with intensity level \code{formant_dynrange} below the maximum
intensity. Default is \code{30}. If set to \code{0}, all formants are shown.}

\item{formantrange}{Vector of two integers giving the frequency range to be
used for producing formant plots. Default is \code{c(0,5500)}.}

\item{type}{String giving the type of formant plot to produce;
default is \code{speckle} (a point plot), the only other option is \code{draw} (a line
plot).}

\item{ind}{Integer indexing waveform relative to other plot components.
Default is \code{NULL}.}

\item{nframe}{Integer giving the number of plot components. Default is \code{NULL}.}
}
\description{
Function for plotting formant objects called by \link{praatpicture}. Instead of
using this function directly, just use
\code{praatpicture('my_sound_file', frames='formant')}.
}
\examples{
#dont use directly
datapath <- system.file('extdata', package='praatpicture')
praatpicture(paste0(datapath, '/1.wav'), frames='formant')
}
