% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specplot.R
\name{specplot}
\alias{specplot}
\title{Plot spectrogram}
\usage{
specplot(
  sig,
  sr,
  t,
  start,
  end,
  tfrom0 = TRUE,
  freqrange = c(0, 5000),
  windowlength = 0.005,
  dynamicrange = 60,
  timestep = 1000,
  windowshape = "Gaussian",
  tgbool = FALSE,
  lines = NULL,
  ind = NULL,
  nframe = NULL
)
}
\arguments{
\item{sig}{Numeric vector corresponding to a sound signal.}

\item{sr}{Integer giving the sampling rate of the signal.}

\item{t}{Numeric vector giving times corresponding to the signal.}

\item{start}{Start time (in seconds) of desired plotted area.}

\item{end}{End time (in seconds) of desired plotted area.}

\item{tfrom0}{Logical; should time on the x-axis run from 0 or from the
original time? Default is \code{TRUE}.}

\item{freqrange}{Vector of two integers giving the frequency range to be
used for plotting spectrograms. Default is \code{c(0,5000)}.}

\item{windowlength}{Window length in seconds for generating spectrograms.
Default is \code{0.005}.}

\item{dynamicrange}{Dynamic range in dB for generating spectrograms. The
maximum intensity minus \code{dynamicrange} will all be printed in white. Default
is \code{50}.}

\item{timestep}{How many time steps should be calculated for spectrograms?
Default is \code{1000}. Note that this takes a while to plot, so for fiddling with
plotting parameters it is a good idea to choose a smaller value.}

\item{windowshape}{String giving the name of the window shape to be applied
to the signal when generating spectrograms. Default is \code{Gaussian}; other
options are \code{square}, \code{Hamming}, \code{Bartlett}, or \code{Hanning}.}

\item{tgbool}{Logical; should dotted lines be plotted corresponding to
locations in a TextGrid? Default is \code{FALSE}.}

\item{lines}{Numeric vector giving locations in seconds of locations from
a TextGrid to be plotted with dotted lines. Default is \code{NULL}.}

\item{ind}{Integer indexing waveform relative to other plot components.
Default is \code{NULL}.}

\item{nframe}{Integer giving the number of plot components. Default is \code{NULL}.}
}
\description{
Function for plotting spectrograms called by \link{praatpicture}. Instead of using
this function directly, just use
\code{praatpicture('my_sound_file', frames='spectrogram')}.
}
\examples{
#dont use directly
datapath <- system.file('extdata', package='praatpicture')
praatpicture(paste0(datapath, '/1.wav'), frames='spectrogram')
}
