% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waveplot.R
\name{waveplot}
\alias{waveplot}
\title{Plot waveform}
\usage{
waveplot(
  sig,
  bit,
  t,
  nchan = 1,
  color = "black",
  tgbool = FALSE,
  lines = NULL,
  focusTierColor = "black",
  focusTierLineType = "dotted",
  ind = NULL,
  nframe = NULL,
  rect_comp = NULL,
  arr_comp = NULL,
  annot_comp = NULL,
  draw_rectangle = NULL,
  draw_arrow = NULL,
  annotate = NULL,
  channelNames = FALSE,
  cn = NULL,
  start_end_only = TRUE,
  min_max_only = TRUE
)
}
\arguments{
\item{sig}{Numeric vector corresponding to a sound signal.}

\item{bit}{Numeric; will generally be grabbed from a loaded \code{WaveMC} object.}

\item{t}{Numeric vector giving times corresponding to the signal.}

\item{nchan}{Numeric; how many channels will be plotted? Default is \code{1}.}

\item{color}{String giving the name of the color to be used for plotting
the waveform. Default is \code{'black'}.}

\item{tgbool}{Logical; should dotted lines be plotted corresponding to
locations in a TextGrid? Default is \code{FALSE}.}

\item{lines}{Numeric vector giving locations in seconds of locations from
a TextGrid to be plotted with dotted lines. Default is \code{NULL}.}

\item{focusTierColor}{String or vector of strings giving the color(s) to
use for plotting focus tier lines. If multiple tiers are focused, a vector
of the same length can be passed, and the nth tier will be plotted in the
nth color. Default is \code{'black'}.}

\item{focusTierLineType}{String or vector of strings giving the line
type(s) for plotting focus tier lines. If multiple tiers are focused, a
vector of the same length can be passed, and the nth tier will be plotted in
the nth line type. Default is \code{'dotted'}.}

\item{ind}{Integer indexing waveform relative to other plot components.
Default is \code{NULL}.}

\item{nframe}{Integer giving the number of plot components. Default is \code{NULL}.}

\item{rect_comp}{Vector of strings or numbers giving plot components to draw
rectangles on. Default is \code{NULL}.}

\item{arr_comp}{Vector of strings of numbers giving plot components to draw
arrows on. Default is \code{NULL}.}

\item{annot_comp}{Vector of strings of numbers giving plot components to
annotate. Default is \code{NULL}.}

\item{draw_rectangle}{List of arguments for drawing rectangles passed from
\code{praatpicture()}. Default is \code{NULL}.}

\item{draw_arrow}{List of arguments for drawing arrows passed from
\code{praatpicture()}. Default is \code{NULL}.}

\item{annotate}{List of arguments for annotating passed from
\code{praatpicture()}. Default is \code{NULL}.}

\item{channelNames}{Logical; should names of audio channels be printed on
the y-axis? Default is \code{FALSE}.}

\item{cn}{Vector of strings with channel names to be printed on the y-axis
if \code{channelNames} is \code{TRUE}.}

\item{start_end_only}{Logical; should there only be ticks on the x-axis
for start and end times? Default is \code{TRUE}.}

\item{min_max_only}{Logical; should only minimum and maximum values be given
on the y-axis? Default is \code{TRUE}. Can also be a logical vector if some but
not all plot components should have minimum and maximum values on the y-axis.
Ignored for TextGrid component.}
}
\description{
Function for plotting waveforms called by \link{praatpicture}. Instead of using
this function directly, just use
\code{praatpicture('my_sound_file', frames='sound')}.
}
\examples{
# Don't use directly
datapath <- system.file('extdata', package='praatpicture')
soundFile <- paste0(datapath, '/1.wav')
praatpicture(soundFile, frames='sound')
}
